Renamed the main directory.
