+++ SigfolioWallet/SigfolioWallet.csproj
