+++ SigfolioWallet/Views/HomeView.xaml
@@ -1,0 +1,15 @@
+﻿<Page
+    x:Class="SigfolioWallet.HomeView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:SigfolioWallet"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <TextBlock Text="Home" TextWrapping="Wrap" VerticalAlignment="Center" FontSize="48" HorizontalAlignment="Center"/>
+
+    </Grid>
+</Page>
