--- StellarNowWallet/Views/HistoryView.xaml
@@ -1,15 +1,0 @@
-﻿<Page
-    x:Class="SigfolioWallet.Views.HistoryView"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:SigfolioWallet.Views"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-
-    <Grid>
-        <TextBlock Text="History" TextWrapping="Wrap" VerticalAlignment="Center" FontSize="48" HorizontalAlignment="Center"/>
-
-    </Grid>
-</Page>
